BACKPORT: KVM: arm/arm64: Remove pmc->bitmask
We currently use pmc->bitmask to determine the width of the pmc - however
it's superfluous as the pmc index already describes if the pmc is a cycle
counter or event counter. The architecture clearly describes the widths of
these counters.

Let's remove the bitmask to simplify the code.

This patch is a part of KVM PMU patchset to fix the overflow issue
(pmu-mem-access, pmu-chain-promotion, pmu-overflow-interrupt) in CentOS 8.0
This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=v5.10-rc6&id=218907cbc2b82419c70180610163c987d4764b27

Signed-off-by: Andrew Murray <andrew.murray@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>