BACKPORT: usb: add a hcd_uses_dma helper
The USB buffer allocation code is the only place in the usb core (and in
fact the whole kernel) that uses is_device_dma_capable, while the URB
mapping code uses the uses_dma flag in struct usb_bus.  Switch the buffer
allocation to use the uses_dma flag used by the rest of the USB code,
and create a helper in hcd.h that checks this flag as well as the
CONFIG_HAS_DMA to simplify the caller a bit.

If MSI is supported, somehow the check of is_device_dma_capable caused
issue, and buffer could not be created. This patch switched the buffer
allocation of USB code to use another flag and helped to resolve the
issue.
This patch is a part of usb patch set to fix the crash in CentOS 8.0.
This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=v5.6-rc1&id=edfbcb321faf07ca970e4191abe061deeb7d3788

Signed-off-by: Christoph Hellwig <hch@lst.de>
Link: https://lore.kernel.org/r/20190811080520.21712-3-hch@lst.de
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>