BACKPORT: ACPI/PPTT: Add support for ACPI 6.3 thread flag
ACPI 6.3 adds a flag to the CPU node to indicate whether
the given PE is a thread. Add a function to return that
information for a given linux logical CPU.

This patch is the first patch of the series of two patches which are
backported from upstream in order to fix the issue with kernel
topology that parses the ACPI Processor ID incorrectly seen in cpu
topology sysfs (/sys/device/system/cpu/cpuX/topology/core_id).

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=bbd1b70639f785a970d998f35155c713f975e3ac

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>
Reviewed-by: Sudeep Holla <sudeep.holla@arm.com>
Reviewed-by: Robert Richter <rrichter@marvell.com>
Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Will Deacon <will@kernel.org>
Signed-off-by: Chuong Tran <chuong.tran@amperecomputing.com>