NOUPSTREAM: centos: Modify emag_defconfig to support more NUMA nodes
Increase the value of NODES_SHIFT to 6 to allow the emag kernel to
support a maximum of 64 NUMA nodes instead of just 8. This will help
with the configuration where we add memory only NUMA nodes for
32-bit memory regions for each socket, CCIX attached memory or
NVDIMM.

The value of 64 brings the number of maximum NUMA nodes to match x86
defconfig and also reflects discussions with Redhat and Oracle. This
value has also been proposed for the upstream 5.10 kernel and is
being discussed.

Signed-off-by: Vanshidhar Konda <vkonda@amperecomputing.com>